+++ offer-service/demo/src/main/java/com/warehouseManagement/demo/dto/CartItemDTO.java
@@ -2,11 +2,13 @@ package com.warehouseManagement.demo.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class CartItemDTO {
     private int offerId;
     private String name;
     private int quantity;
-    private double unitPrice;
-    private double lineTotal;
+    private BigDecimal unitPrice;
+    private BigDecimal lineTotal;
 }
