+++ offer-service/demo/src/main/java/com/warehouseManagement/demo/services/OfferService.java
@@ -10,6 +10,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 
@@ -51,7 +52,7 @@ public class OfferService {
     }
 
     @Transactional
-    public void changeOffer(int userId, int offerId, float price, int availableQuantity, int minimalQuantity) {
+    public void changeOffer(int userId, int offerId, BigDecimal price, int availableQuantity, int minimalQuantity) {
         Wholesaler wholesaler = wholesalerRepository.findByUser_Id(userId);
         if(!offerRepository.existsByIdAndWholesaler(offerId,wholesaler)){
             throw new RuntimeException("Offer does not exist or you are not the seller of that offer");
